requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.0
tqdm==4.66.4